# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Crop Digital Twin API
  description: Local farm management for South China rice; simulation is pending.
  version: 0.3.0
paths:
...
  /api/v1/plots:
    get:
      tags:
      - farm
      summary: List Plots
      description: Page saved plots in newest-first order.
      operationId: listPlots
      security:
      - APIKeyCookie: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_PlotResponse_'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - farm
      summary: Create Plot
      description: Save a manual plot with its original area and position.
      operationId: createPlot
      security:
      - APIKeyCookie: []
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlotCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlotResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/plots/{plot_id}:
    get:
      tags:
      - farm
      summary: Get Plot
      description: Read one plot, or return a documented 404.
      operationId: getPlot
      security:
      - APIKeyCookie: []
      parameters:
      - name: plot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plot Id
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlotResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seasons:
    get:
      tags:
      - farm
      summary: List Seasons
      description: Page rice seasons belonging to one plot.
      operationId: listSeasons
      security:
      - APIKeyCookie: []
      parameters:
      - name: plot_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Plot Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_SeasonResponse_'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - farm
      summary: Create Season
      description: Establish a non-overlapping rice season.
      operationId: createSeason
      security:
      - APIKeyCookie: []
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeasonCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/seasons/{season_id}/close:
    post:
      tags:
      - farm
      summary: Close Season
      description: Confirm season completion after all still-current operations.
      operationId: closeSeason
      security:
      - APIKeyCookie: []
      parameters:
      - name: season_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Season Id
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeasonClose'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/management-events:
    get:
      tags:
      - farm
      summary: List Events
      description: Page the current operation versions, or include retained history.
      operationId: listManagementEvents
      security:
      - APIKeyCookie: []
      parameters:
      - name: season_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Season Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: include_history
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include History
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_EventResponse_'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - farm
      summary: Create Event
      description: Save an operation with original and standard-unit quantities.
      operationId: createManagementEvent
      security:
      - APIKeyCookie: []
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/management-events/{event_id}/corrections:
    post:
      tags:
      - farm
      summary: Correct Event
      description: Append a corrected operation without overwriting its predecessor.
      operationId: correctManagementEvent
      security:
      - APIKeyCookie: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Event Id
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCorrection'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: ErrorDetail
      description: Safe business/storage error, with no credentials or SQL parameters.
    ErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Documented stable error envelope.
...
    EventCreate:
      properties:
        event_type:
          type: string
          enum:
          - sowing
          - transplanting
          - irrigation
          - fertilization
          - inspection
          - harvest
          title: Event Type
        occurred_on:
          type: string
          format: date
          title: Occurred On
        quantity:
          anyOf:
          - type: number
            maximum: 1000000.0
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d{0,6}0*$
          - type: 'null'
          title: Quantity
        unit:
          anyOf:
          - type: string
            enum:
            - mm
            - m3
            - kg/mu
            - kg/ha
          - type: 'null'
          title: Unit
        material_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Material Name
        notes:
          type: string
          maxLength: 2000
          title: Notes
          default: ''
        season_id:
          type: string
          format: uuid
          title: Season Id
      additionalProperties: false
      type: object
      required:
      - event_type
      - occurred_on
      - season_id
      title: EventCreate
      description: New operation in a specified season.
    EventResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        season_id:
          type: string
          format: uuid
          title: Season Id
        event_type:
          type: string
          enum:
          - sowing
          - transplanting
          - irrigation
          - fertilization
          - inspection
          - harvest
          title: Event Type
        occurred_on:
          type: string
          format: date
          title: Occurred On
        quantity:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Quantity
        unit:
          anyOf:
          - type: string
            enum:
            - mm
            - m3
            - kg/mu
            - kg/ha
          - type: 'null'
          title: Unit
        normalized_quantity:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Normalized Quantity
        normalized_unit:
          anyOf:
          - type: string
            enum:
            - mm
            - kg/ha
          - type: 'null'
          title: Normalized Unit
        material_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Material Name
        notes:
          type: string
          title: Notes
        created_at:
          type: string
          format: date-time
          title: Created At
        revision:
          type: integer
          title: Revision
        replaces_event_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Replaces Event Id
        correction_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Correction Reason
        is_current:
          type: boolean
          title: Is Current
      type: object
      required:
      - id
      - season_id
      - event_type
      - occurred_on
      - quantity
      - unit
      - normalized_quantity
      - normalized_unit
      - material_name
      - notes
      - created_at
      - revision
      - replaces_event_id
      - correction_reason
      - is_current
      title: EventResponse
      description: Original/standard quantities and a traceable operation revision.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PageResponse_EventResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EventResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PageResponse[EventResponse]
    PageResponse_PlotResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PlotResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PageResponse[PlotResponse]
    PageResponse_SeasonResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SeasonResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PageResponse[SeasonResponse]
...
    PlotCreate:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        area_mu:
          anyOf:
          - type: number
            maximum: 1000000.0
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d{0,6}0*$
          title: Area Mu
        latitude:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Latitude
        longitude:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Longitude
      additionalProperties: false
      type: object
      required:
      - name
      - area_mu
      - latitude
      - longitude
      title: PlotCreate
      description: Reported area in mu and a WGS84 point.
    PlotResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        name:
          type: string
          title: Name
        area_mu:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Area Mu
        area_ha:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Area Ha
        latitude:
          type: number
          title: Latitude
        longitude:
          type: number
          title: Longitude
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - organization_id
      - name
      - area_mu
      - area_ha
      - latitude
      - longitude
      - created_at
      title: PlotResponse
      description: Reported plot facts and derived hectares.
    SeasonClose:
      properties:
        end_date:
          type: string
          format: date
          title: End Date
      additionalProperties: false
      type: object
      required:
      - end_date
      title: SeasonClose
      description: Confirmed season end date, not a predicted harvest date.
    SeasonCreate:
      properties:
        plot_id:
          type: string
          format: uuid
          title: Plot Id
        start_date:
          type: string
          format: date
          title: Start Date
        establishment_method:
          type: string
          enum:
          - direct_sowing
          - transplanting
          title: Establishment Method
        variety_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Variety Name
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
      additionalProperties: false
      type: object
      required:
      - plot_id
      - start_date
      - establishment_method
      title: SeasonCreate
      description: Rice establishment; absent variety is recorded as unknown.
    SeasonResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        plot_id:
          type: string
          format: uuid
          title: Plot Id
        crop_code:
          type: string
          const: rice
          title: Crop Code
        start_date:
          type: string
          format: date
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        establishment_method:
          type: string
          enum:
          - direct_sowing
          - transplanting
          title: Establishment Method
        variety_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Variety Name
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - plot_id
      - crop_code
      - start_date
      - end_date
      - establishment_method
      - variety_name
      - created_at
      title: SeasonResponse
      description: The rice season state and establishment facts.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: crop_twin_session